Extract `RemoteNearestStop` type to its own file
